+++ lib/templates.php
@@ -17,7 +17,7 @@ function include_template_function($template_path)
                 return $theme_file;
             }
             return plugin_dir_path(dirname(__FILE__)) . '/lib/single-proceedings.php';
-        } elseif (is_archive() && !is_search()) {
+        } elseif (is_archive() && !is_search() && !is_author()) {
             // checks if the file exists in the theme first,
             // otherwise serve the file from the plugin
             if ($theme_file = locate_template(array ('archive-proceedings.php'))) {